NeuroBridge/rag_brain/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
faiss-cpu
sentence-transformers
pypdf
python-docx
pandas
torch
